Split cli to SvgXmlToImageVectorCommand and IconPackCommand, bump dependencies
