+++ cli/src/test/kotlin/io/github/composegears/valkyrie/cli/IntegrationTest.kt
@@ -6,12 +6,14 @@ import java.nio.file.Path
 import kotlin.io.path.createTempDirectory
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.MethodSource
 
 class IntegrationTest {
 
     @ParameterizedTest
+    @Disabled
     @MethodSource("testMatrix")
     fun exportAndImport(arg: Pair<Boolean, OutputFormat>) {
         val (useCli, outputFormat) = arg
