+++ components/generator/iconpack/src/main/kotlin/io/github/composegears/valkyrie/generator/iconpack/IconPackGenerator.kt
@@ -3,6 +3,7 @@ package io.github.composegears.valkyrie.generator.iconpack
 data class IconPackGeneratorConfig(
     val packageName: String,
     val iconPackName: String,
+    // TODO: rename in all places to nestedPacks
     val subPacks: List<String>,
     val useExplicitMode: Boolean,
     val indentSize: Int,
