+++ gradle/libs.versions.toml
@@ -20,8 +20,8 @@ tiamat = "io.github.composegears:tiamat:1.2.0"
 
 xpp3 = "org.ogce:xpp3:1.1.6"
 
-clikt = "com.github.ajalt.clikt:clikt:4.4.0"
-r8 = "com.android.tools:r8:8.2.47"
+clikt = "com.github.ajalt.clikt:clikt:5.0.1"
+r8 = "com.android.tools:r8:8.5.35"
 
 junit5-jupiter = "org.junit.jupiter:junit-jupiter:5.11.3"
 junit-launcher = { module = "org.junit.platform:junit-platform-launcher" }
